Changes 13-01-13 @ 19:05
> Finished annotation of `disabilityEditor::disabilityEditor()` working
off previous thougths as mentioned in the
b100a1475f833882dca9b0346e9b5bea5fb27f16 commit notes. Should be noted
that comments 1 & 2 in this function are copied from the comments in
`seatdisplay::seatdisplay()` word for word, and may need to be changed,
but these changes can be merged into master.

> Deleted "test_unfin_codeBRANCH" as I have now tested the branch system
and it works with my GitHub app.